+++ src/main/java/com/dsmpear/main/exceptions/UserIsAlreadyRegisteredException.java
@@ -1,0 +1,7 @@
+package com.dsmpear.main.exceptions;
+
+public class UserIsAlreadyRegisteredException extends RuntimeException {
+    public UserIsAlreadyRegisteredException() {
+        super("User is already registered Exception");
+    }
+}
