+++ src/main/java/com/dsmpear/main/service/user/UserService.java
@@ -1,4 +1,7 @@
 package com.dsmpear.main.service.user;
 
+import com.dsmpear.main.payload.request.RegisterRequest;
+
 public interface UserService {
+    public void register(RegisterRequest request);
 }
