+++ src/test/java/com/dsmpear/main/QuestionControllerTest.java
@@ -1,0 +1,47 @@
+package com.dsmpear.main;
+
+import com.dsmpear.main.payload.request.QuestionRequest;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
+import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+class QuestionControllerTest {
+
+
+    @Autowired
+    private WebApplicationContext context;
+
+    private MockMvc mvc;
+
+    @BeforeEach
+    public void setup() {
+        mvc = MockMvcBuilders
+              .webAppContextSetup(context)
+                .build();
+    }
+    @Test
+    void test() throws Exception {
+
+        QuestionRequest request = new QuestionRequest("asd@asd.asd", "asdasdasds");
+
+        mvc.perform(post("/question")
+                .content(new ObjectMapper()
+                        .writeValueAsString(request))
+                .contentType(MediaType.APPLICATION_JSON_VALUE))
+                .andExpect(status().isOk()).andDo(print());
+    }
+
+}
