--- src/test/java/com/dsmpear/main/domain/TokenControllerTest.java
@@ -1,69 +1,0 @@
-package com.dsmpear.main.domain;
-
-import com.dsmpear.main.entity.refreshtoken.RefreshTokenRepository;
-import com.dsmpear.main.entity.user.User;
-import com.dsmpear.main.entity.user.UserRepository;
-import com.dsmpear.main.payload.request.SignInRequest;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class TokenControllerTest {
-    @Autowired
-    private WebApplicationContext context;
-
-    private MockMvc mvc;
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @Autowired
-    private RefreshTokenRepository tokenRepository;
-
-    @Autowired
-    private PasswordEncoder passwordEncoder;
-
-    @BeforeEach
-    public void setup() {
-        mvc = MockMvcBuilders
-                .webAppContextSetup(context)
-                .build();
-
-        userRepository.save(
-                User.builder()
-                .email("aaaa@dsm.hs.kr")
-                .password(passwordEncoder.encode("1111"))
-                .build()
-        );
-    }
-
-    @AfterEach
-    public void exit() {
-        userRepository.deleteAll();
-        tokenRepository.deleteAll();
-    }
-
-    @Test
-    void test() throws Exception {
-        SignInRequest signInRequest = new SignInRequest("aaaa@dsm.hs.kr", "1111");
-
-        mvc.perform(post("/auth").content(new ObjectMapper().writeValueAsString(signInRequest))
-                .contentType(MediaType.APPLICATION_JSON_VALUE))
-                .andExpect(status().isOk()).andDo(print());
-    }
-}
